backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
anthropic>=0.18.1
yt-dlp>=2024.1.1
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
aiosqlite>=0.19.0
email-validator>=2.0.0
# openai-whisper - install separately if needed for transcription
